requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.0
python-dotenv==1.0.0
PyMySQL==1.0.3